feat: move into `harpc-client`
