+++ apps/hash-graph/libs/api/src/rpc/account.rs
@@ -465,6 +465,11 @@ where
 
 // TODO: delegate, this can be auto generated by the `harpc` crate
 
+// TODO: move this into `harpc_client`
+#[derive(Debug, derive_more::Display, derive_more::Error)]
+#[display("unknown error: {_0:?}")]
+struct UnknownError(#[error(not(source))] serde_json::Value);
+
 // TODO: this can be auto generated by the `harpc` crate
 pub struct AccountClient;
