+++ libs/@local/harpc/client/Cargo.toml
@@ -11,14 +11,17 @@ publish.workspace = true
 [dependencies]
 # Public workspace dependencies
 harpc-tower = { workspace = true, public = true }
+harpc-service = { workspace = true, public = true }
 
 # Public third-party dependencies
 tower = { workspace = true, public = true }
+serde-value = {version = "0.7.0", public = true }
 
 # Private workspace dependencies
-error-stack = { workspace = true }
+error-stack = { workspace = true, features = ["unstable", "futures"] }
 harpc-codec = { workspace = true }
 harpc-net = { workspace = true }
+harpc-types = { workspace = true }
 
 # Private third-party dependencies
 bytes = { workspace = true }
@@ -27,6 +30,8 @@ futures = { workspace = true }
 multiaddr = { workspace = true }
 thiserror = { workspace = true }
 tokio-util = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+derive_more = { version = "1.0.0", features = ["display"] }
 
 [lints]
 workspace = true
