+++ libs/@local/harpc/client/src/lib.rs
@@ -2,7 +2,9 @@
 
 extern crate alloc;
 
+pub mod common;
 pub mod connection;
+pub mod error;
 
 use alloc::sync::Arc;
